--- deprecated_tests/google-site-search/fetch.spec.ts
@@ -1,31 +1,0 @@
-import test from "ava";
-// import casual from "casual";
-
-import { GoogleConnector } from "../../src/google-site-search/fetch";
-
-// test("`connect` should fail without any env vars", async (t) => {
-//   const SS = await connect(casual.url);
-//   t.falsy(SS);
-// });
-
-// test("`connect` should fail without all env vars", async (t) => {
-//   process.env.SEARCH_URL = casual.url;
-//   process.env.SEARCH_KEY = casual.word;
-
-//   const SS = await connect(casual.url);
-//   t.falsy(SS);
-// });
-
-// test("`connect` should return true if proper env vars", async (t) => {
-//   process.env.SEARCH_URL = casual.url;
-//   process.env.SEARCH_KEY = casual.word;
-//   process.env.SEARCH_CX = casual.word;
-
-//   const SS = await connect(casual.url);
-//   t.truthy(SS);
-// });
-
-test("`GoogleConnector` should expose get function", async (t) => {
-  const testFetcher = new GoogleConnector();
-  t.truthy(testFetcher.get);
-});
