--- deprecated_tests/rock/models/finances/resolver.spec.ts
@@ -1,46 +1,0 @@
-import test from "ava";
-// import Resolver from "../../../../src/rock/models/campuses/resolver";
-
-// const sampleData = {
-//   transaction: {
-//     Id: "123ABC",
-//     Amount: "25.00",
-//     AccountId: "12345",
-//   },
-// };
-
-// test("`TransactionDetail` has amount and account.", t => {
-//   t.pass();
-// });
-
-// test("`ScheduledTransaction` has start and end date.", t => {
-//   t.pass();
-// });
-
-// test("`ScheduledTransaction` has required meta data.", t => {
-//   t.pass();
-// });
-
-// test("`Transaction` has payment details.", t => {
-//   t.pass();
-// });
-
-// test("`FinancialAccount` is a valid account for contributions.", t => {
-//   t.pass();
-// });
-
-// test("`PaymentDetail` has detailed information about a single payment.", t => {
-//   t.pass();
-// });
-
-// test("`SavedPayment` has the required information to repeat transactions.", t => {
-//   t.pass();
-// });
-
-test.todo("`TransactionDetail` has amount and account.");
-test.todo("`ScheduledTransaction` has start and end date.");
-test.todo("`ScheduledTransaction` has required meta data.");
-test.todo("`Transaction` has payment details.");
-test.todo("`FinancialAccount` is a valid account for contributions.");
-test.todo("`PaymentDetail` has detailed information about a single payment.");
-test.todo("`SavedPayment` has the required information to repeat transactions.");
