+++ src/apollos/__tests__/mongo.spec.ts
@@ -1,0 +1,61 @@
+
+import { Schema } from "mongoose";
+
+import { MongoConnector, connect } from "../mongo";
+
+describe("connect", () => {
+  it("\'connect\' should allow for a connection status to be returned", async () => {
+    const status = await connect("fake address");
+    expect(status).toBeFalsy();
+  });
+});
+
+
+describe("MongoConnector", () => {
+
+  let testModel;
+  beforeEach(async () => {
+    if (testModel) return;
+    await connect("fake address");
+    const schema = { _id: String };
+    testModel = new MongoConnector("test", schema);
+  });
+
+
+  it(" should expose the db connection", () => {
+    expect(testModel.db).toBeTruthy();
+  });
+
+  it("should create a pluralized collection based off the model name", () => {
+    expect(testModel.model.collection.collectionName).toEqual("tests");
+  });
+
+  it("should create a model based on the name passed", () => {
+    expect(testModel.model.modelName).toEqual("test");
+  });
+
+  it("should use the schema passed as the second argument", () => {
+    const { schema } = testModel.model;
+
+    expect(schema instanceof Schema).toBeTruthy();
+    expect(schema.paths._id).toBeTruthy();
+    expect(schema.paths._id.instance).toEqual("String");
+  });
+
+  it("findOne should be a sugared passthrough to the models findOne", async () => {
+    const oldFindOne = testModel.model.findOne;
+
+    testModel.model.findOne = function mockedFindOne(...args) {
+      expect(args[0]).toEqual("test");
+      return new Promise((r) => r([1]));
+    };
+
+    const tests = await testModel.findOne("test");
+    expect(tests[0]).toEqual(1);
+
+    testModel.model.findOne = oldFindOne;
+
+  });
+
+
+});
