+++ src/rock/models/finances/__test__/resolver.spec.ts
@@ -1,0 +1,46 @@
+// import Resolver from "../resolver";
+
+// const sampleData = {
+//   transaction: {
+//     Id: "123ABC",
+//     Amount: "25.00",
+//     AccountId: "12345",
+//   },
+// };
+
+// test("`TransactionDetail` has amount and account.", t => {
+//   t.pass();
+// });
+
+// test("`ScheduledTransaction` has start and end date.", t => {
+//   t.pass();
+// });
+
+// test("`ScheduledTransaction` has required meta data.", t => {
+//   t.pass();
+// });
+
+// test("`Transaction` has payment details.", t => {
+//   t.pass();
+// });
+
+// test("`FinancialAccount` is a valid account for contributions.", t => {
+//   t.pass();
+// });
+
+// test("`PaymentDetail` has detailed information about a single payment.", t => {
+//   t.pass();
+// });
+
+// test("`SavedPayment` has the required information to repeat transactions.", t => {
+//   t.pass();
+// });
+
+// tslint:disable
+xit("`TransactionDetail` has amount and account.", () => {});
+xit("`ScheduledTransaction` has start and end date.", () => {});
+xit("`ScheduledTransaction` has required meta data.", () => {});
+xit("`Transaction` has payment details.", () => {});
+xit("`FinancialAccount` is a valid account for contributions.", () => {});
+xit("`PaymentDetail` has detailed information about a single payment.", () => {});
+xit("`SavedPayment` has the required information to repeat transactions.", () => {});
